% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{ConfusionMatrix}
\alias{ConfusionMatrix}
\title{\code{ConfusionMatrix}}
\usage{
ConfusionMatrix(obj, subset = obj$subset, weights = obj$weights)
}
\arguments{
\item{obj}{A model with an outcome variable.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process or the name of a variable in \code{data}. It
may not be an expression.}

\item{weights}{An optional vector of sampling weights or the
name of a variable in \code{data}. It may not be an expression.}
}
\description{
\code{ConfusionMatrix}
}
\details{
Produces a confusion matrix for a trained model showing the proportion
of observed values that take the same values as the predicted values. Where the outcome variable
in the model is not a factor and not a count, observed and predicted values are assigned to buckets.
}

