% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{ConfusionMatrixFromVariablesNumeric}
\alias{ConfusionMatrixFromVariablesNumeric}
\title{\code{ConfusionMatrixFromVariablesNumeric}}
\usage{
ConfusionMatrixFromVariablesNumeric(observed, predicted, subset = NULL,
  weights = NULL)
}
\arguments{
\item{observed}{A \code{numeric}.}

\item{predicted}{A \code{numeric}.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or the name of a variable in \code{data}. It
may not be an expression.}

\item{weights}{An optional vector of sampling weights or the name
name of a variable in \code{data}. It may not be an expression.}
}
\description{
\code{ConfusionMatrixFromVariablesNumeric}
}
\details{
A contingency table showing the observed versus predicted values
where predicted values are assigned their closest observed value.
}
