% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{cooks.distance.Regression}
\alias{cooks.distance.Regression}
\title{\code{cooks.distance.Regression}}
\usage{
\method{cooks.distance}{Regression}(model, ...)
}
\arguments{
\item{model}{Regression model.}

\item{...}{Other arugments.}
}
\description{
\code{cooks.distance.Regression}
}
