% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionplots.R
\name{EffectsPlot}
\alias{EffectsPlot}
\title{Produce an effects plot of a regression model}
\usage{
EffectsPlot(model, max.factor.label.length = NULL, y.axis.title = NULL)
}
\arguments{
\item{model}{A model of class \code{Regression}}

\item{max.factor.label.length}{The maximum length of any factor label, as plotted on the
x-axis for categorical variables. Longer labels are truncated, possibly with an integer
suffix to ensure uniquueness.}

\item{y.axis.title}{The title of the y-axis for each plot. Defaults to the outcome label.}
}
\description{
Produce an effects plot of a regression model
}
