% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{FitRegression}
\alias{FitRegression}
\title{\code{FitRegression}}
\usage{
FitRegression(
  .formula,
  .estimation.data,
  .weights,
  type,
  robust.se,
  outlier.prop.to.remove,
  seed = 12321,
  dummy.processed.data = NULL,
  ...
)
}
\arguments{
\item{.formula}{An object of class \code{\link{formula}} (or one that can be
coerced to that class): a symbolic description of the model to be fitted.
The details of type specification are given under \sQuote{Details}.}

\item{.estimation.data}{A \code{\link{data.frame}}.}

\item{.weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{type}{See \link{Regression}.}

\item{robust.se}{Computes standard errors that are robust to violations of
the assumption of constant variance, using the HC1 (degrees of freedom)
modification of White's (1980) estimator (Long and Ervin, 2000). This parameter is ignored
if weights are applied (as weights already employ a sandwich estimator).}

\item{outlier.prop.to.remove}{A single numeric value that determines the percentage of data points to remove from the
analysis. The data points removed correspond to those in the proportion with the largest residuals.
A value of 0 or NULL would denote no points are removed. A value x, with 0 < x < 0.5 (not inclusive) would
denote that a percentage between none and 50\% of the data points are removed.}

\item{seed}{A integer seed to generate the surrogate residuals.}

\item{dummy.processed.data}{A list containing the output from \code{\link[flipData]{EstimationData}} when the dummy
adjustment for missing data is used. It is only needed when outliers are removed with dummy adjustment. In that
situation, the estimation data needs to be recalibrated in case any adjustment variables are redundant or singular
when outliers are removed from the data.}

\item{...}{Arguments to the wrapped functions.}
}
\description{
Fits a regression model.
}
