% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{FixVarianceCovarianceMatrix}
\alias{FixVarianceCovarianceMatrix}
\title{FixVarianceCovarianceMatrix}
\usage{
FixVarianceCovarianceMatrix(x, min.eigenvalue = 1e-12)
}
\arguments{
\item{x}{A variance-covariance matrix of parameter estimates.}

\item{min.eigenvalue}{Minimm acceptable eigenvalue.}
}
\description{
Makes some adjustments to permit a covariance-marix to be inverted, if required.
}
\details{
Sandwich and sandwich-like standard errors can result uninvertable
covariance matrices (e.g., if a parameter represents a sub-group, and the sub-group has no
residual variance). This function checks to see if there are any eigenvalues less than \code{min.eigenvalue},
which defaults to 1e-12. If there are, an attempt is made to guess the offending variances, and they are multiplied by 1.01.
}
