% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R, R/regressiondiagnostics.R
\name{GoodnessOfFit}
\alias{GoodnessOfFit}
\alias{GoodnessOfFit.Regression}
\alias{GoodnessOfFit.default}
\title{Object Goodness-of-Fit
\code{GoodnessOfFit} summary is a generic function used to produce result summaries of the
results of the model object. The function invokes particular \code{\link{methods}}
which depend on the \code{\link{class}} of the first argument.}
\usage{
GoodnessOfFit(object, digits = max(3L, getOption("digits") - 3L), ...)

\method{GoodnessOfFit}{default}(object, digits = max(3L, getOption("digits") -
  3L), ...)

\method{GoodnessOfFit}{Regression}(object, digits = max(3L,
  getOption("digits") - 3L), ...)
}
\arguments{
\item{object}{An object for which a summary is desired..}

\item{digits}{Minimal number of significant digits, see \code{\link{print.default}}.}

\item{...}{Additional arguments affecting the goodness-of-fit displayed.
#@examples

# linear regression
x <- rnorm(10)
y <- rnorm(10) + x
mod <- lm(y ~ x)
GoodnessOfFit(mod)

# MDS - square
library(smacof)
data(breakfastDissimilarities)
mdsInterval <- smacofSym(breakfastDissimilarities[[4]],
    type = "interval", eps = 1e-12, itmax = 100000)
GoodnessOfFit(mdsInterval)
# MDS - rectangular
data(breakfast)
GoodnessOfFit(smacofRect(breakfast))}
}
\description{
reports the goodness-of-fit of an object.

#' \code{BreuschPagan} Breusch-Pagan test for non-constant variance.
#'
#' @param mod An object of class \code{\link{Regression}}.
#' @param show.warnings Returns a warning if the sample size is less than 4.
#' @details Weights are ignored when conducting the test.
#' @export
BreuschPagan <- function(Regression.object, show.warnings = TRUE)#, var.formula)
{
    if (class(Regression.object)[1] != "Regression")
        return(car::ncvTest(Regression.object))
    #Modified from car::ncvTest, to addess different size of input data.
    subset <- Regression.object$subset
    if(sum(subset) < 4)
    {
        if (show.warnings)
            warning("Sample size too small for Breusch-Pagan test.")
        return(list(ChiSquare = NA, Df = NA, p = 1,
            test = "Breusch-Pagan test of Non-constant Variance"))
    }
    residuals <- residuals(Regression.object)[subset]
    fitted.values <- fitted.values(Regression.object)[subset]
    squared.residuals <- residuals^2
    r.squared <- 1 - var(residuals) / var(fitted.values)
    if (r.squared> 0.9999) {
        Chisq <- 0.0
    }
    else
    {
        U <- squared.residuals / mean(squared.residuals)#mean.squared.error#sum(squared.residuals)
        mod <- lm(U ~ fitted.values)#, subset = Regression.object$subset)
        SS <- anova(mod)$"Sum Sq"
        RegSS <- sum(SS) - SS[length(SS)]
        Chisq <- RegSS/2
    }
    result <- list(#formula = var.formula, formula.name = "Variance",
        ChiSquare = Chisq, Df = 1, p = pchisq(Chisq, 1, lower.tail = FALSE),
        test = "Breusch-Pagan test of Non-constant Variance")
    class(result) <- "chisqTest"
    result
}
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default goodness-of-fit \eqn{R^2}.

\item \code{Regression}: Goodness-of-fit for a Regression object. Computed as the \eqn{R^2} statistic.
With factors, the index value of the factor is used. With unordered factors, this will often be
grieviously-downward biased.
}}

