% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{GoodnessOfFitPlot}
\alias{GoodnessOfFitPlot}
\alias{GoodnessOfFitPlot.Regression}
\title{\code{GoodnessOfFitPlot}}
\usage{
GoodnessOfFitPlot(object, max.points = 1000, ...)

\method{GoodnessOfFitPlot}{Regression}(object, max.points = 1000, ...)
}
\arguments{
\item{object}{An object for which a summary is desired.}

\item{max.points}{The maximum numner of points to plot.}

\item{...}{Additional arguments affecting the goodness-of-fit displayed.}
}
\description{
A generic function used to produce plots illustrating the goodness-of-fit of
the model object.  The function invokes particular \code{\link{methods}}
which depend on the \code{\link{class}} of the first argument.

Reports the goodness-of-fit of an object.
}
\section{Methods (by class)}{
\itemize{
\item \code{Regression}: Goodness-of-fit plot for a Regression object
}}

