% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{GoodnessOfFitPlot.Regression}
\alias{GoodnessOfFitPlot.Regression}
\title{Goodness-of-fit plot for a Regression object}
\usage{
\method{GoodnessOfFitPlot}{Regression}(object, max.points = 1000, ...)
}
\arguments{
\item{object}{An object for which a summary is desired.}

\item{max.points}{The maximum numner of points to plot.}

\item{...}{Additional arguments affecting the goodness-of-fit displayed.}
}
\description{
Goodness-of-fit plot for a Regression object
}
