% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{GrandMean}
\alias{GrandMean}
\title{\code{GrandMean}}
\usage{
GrandMean(model)
}
\arguments{
\item{model}{A 'Regression'  model.}
}
\description{
\code{GrandMean}
}
\details{
Computes the mean of the dependent variable, taking missing values, weights,
and the need to transform into numeric variables into account.
#' @importFrom flipTransformations AsNumeric
}
