% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{infIndexPlot.Regression}
\alias{infIndexPlot.Regression}
\title{\code{infIndexPlot.Regression}}
\usage{
infIndexPlot.Regression(model, ...)
}
\arguments{
\item{model}{A 'Regression'  model.}

\item{...}{Other parameters for \code{infIndexPlot}.}
}
\description{
\code{infIndexPlot.Regression}
}
