% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionsweep.R
\name{LinearRegressionFromCorrelations}
\alias{LinearRegressionFromCorrelations}
\title{\code{LinearRegressionFromCorrelations} Linear regression via a sweep operation n the correlation matrix.}
\usage{
LinearRegressionFromCorrelations(formula, data = NULL, subset = NULL,
  weights = NULL, ...)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}} (or one that can be
coerced to that class): a symbolic description of the model to be fitted.
The details of type specification are given under \sQuote{Details}.}

\item{data}{A \code{\link{data.frame}}.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{...}{Additional argments to be past to  \code{\link[psych]{setCor}}.}
}
\description{
\code{LinearRegressionFromCorrelations} Linear regression via a sweep operation n the correlation matrix.
}
\details{
Estimates regression using sweep operations on the correlation matrix.
The sample size assumed in tests is the smallest pairwise correlation sample size.
}
