% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleimputation.R
\name{MultipleImputationCoefficientTable}
\alias{MultipleImputationCoefficientTable}
\title{\code{MultipleImputationCoefficientTable} Creates the table of coefficients from multiple imputation.}
\usage{
MultipleImputationCoefficientTable(models, large.sample.df = FALSE)
}
\arguments{
\item{models}{A list of models created with multiply-imputed data sets.}

\item{large.sample.df}{Whether to estimate the degrees of freedom for the t-distribution using a
large sample or sample sample approach.}
}
\description{
\code{MultipleImputationCoefficientTable} Creates the table of coefficients from multiple imputation.
}
\references{
J Barnard and DB Rubin (1999) 'Small-sample degrees of freedom with multiple imputation' Biometrika (1999) 86 (4): 948-955.
DB Rubin (1987) 'Multiple Imputation for Nonresponse in Surveys', John Wiley & Sons.
}
