% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{ncvTest.Regression}
\alias{ncvTest.Regression}
\title{#' @importMethodsFrom car Anova
#' @export
setMethod("Anova", "missing", function "<No Object>")
ncvTest.Regression}
\usage{
\method{ncvTest}{Regression}(model, ...)
}
\arguments{
\item{model}{A \code{\link{Regression}} model.}

\item{...}{Additional parameters to \code{\link{ncvTest}}}
}
\description{
#' @importMethodsFrom car Anova
#' @export
setMethod("Anova", "missing", function "<No Object>")
ncvTest.Regression
}

