% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{notValidForCrosstabInteraction}
\alias{notValidForCrosstabInteraction}
\title{notValidForCrosstabInteraction}
\usage{
notValidForCrosstabInteraction(object, method)
}
\arguments{
\item{object}{A Regression object}

\item{method}{The regression method.}
}
\description{
notValidForCrosstabInteraction
}

