% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{notValidForPartial}
\alias{notValidForPartial}
\title{#' @export
coef.Regression <- function(object, ...)
{
    object$coef
}}
\usage{
notValidForPartial(object, method)
}
\arguments{
\item{object}{A Regression object}

\item{The}{regression ethod.}
}
\description{
#' @export
coef.Regression <- function(object, ...)
{
    object$coef
}
}

