% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Observed.FitRegression}
\alias{Observed.FitRegression}
\title{\code{Observed.FitRegression}}
\usage{
\method{Observed}{FitRegression}(x)
}
\arguments{
\item{x}{A 'FitRegression' model.}
}
\description{
Observed values used in fitting a model with an outcome variable.
}
