% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Observed}
\alias{Observed}
\title{\code{observed} Observed values used in fitting a model with an outcome variable.}
\usage{
Observed(x)
}
\arguments{
\item{x}{A 'Regression' model.}
}
\description{
\code{observed} Observed values used in fitting a model with an outcome variable.
}

