% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{predict.Regression}
\alias{predict.Regression}
\title{\code{predict.Regression}}
\usage{
\method{predict}{Regression}(object, newdata = object$model, na.action = na.pass, ...)
}
\arguments{
\item{object}{A \code{Regression} object.}

\item{newdata}{Optionally, a data frame including the variables used to fit the model.
If omitted, the \code{data} supplied to \code{Regression()} is used after any filtering.}

\item{na.action}{Function determining what should be done with missing values in \code{newdata}.
The default is to predict \code{NA}.}

\item{...}{Additional arguments to pass to predict.LDA.}
}
\description{
Predicts a model outcome based on \code{newdata} and a fitted Regression \code{object}.
NA is returned for cases with unfitted factor levels.
}
