% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{PrettyRegressionTable}
\alias{PrettyRegressionTable}
\title{PrettyRegressionTable}
\usage{
PrettyRegressionTable(coefficient.table, t, footer, title = "",
  subtitle = "")
}
\arguments{
\item{coefficient.table}{A table of regression coefficients, standard errors, z or t statistics, and p-values.}

\item{t}{\code{TRUE} if t-statistics, and \code{FALSE} if z-statistics.}

\item{footer}{Text to place in the footer of the table.}

\item{title}{The title for the table.}

\item{subtitle}{Subtitle for the table.}
}
\description{
Creates a pretty formattable table.
}
\references{
This is based on code written by Kenton Russell.
}

