% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.Anova}
\alias{print.Anova}
\title{print.Anova}
\usage{
\method{print}{Anova}(x, ...)
}
\arguments{
\item{x}{An \link{Anova} object.}

\item{...}{Additional parameters to \code{\link{print.Anova}}}
}
\description{
print.Anova
}

