% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{print.ConfusionMatrix}
\alias{print.ConfusionMatrix}
\title{\code{print.ConfusionMatrix}}
\usage{
\method{print}{ConfusionMatrix}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{ConfusionMatrix}}.}

\item{...}{Further arguments, currently unusued.}
}
\description{
\code{print.ConfusionMatrix}
}
\details{
Displays a confusion matrix as a heatmap.
}

