% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{print.DurbinWatson}
\alias{print.DurbinWatson}
\title{\code{print.DurbinWatson}}
\usage{
\method{print}{DurbinWatson}(x, ...)
}
\arguments{
\item{x}{Results from the DurbinWatson test.}

\item{...}{Other arugments.}
}
\description{
\code{print.DurbinWatson}
}
