% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{PrintConfusionMatrix}
\alias{PrintConfusionMatrix}
\title{\code{PrintConfusionMatrix}}
\usage{
PrintConfusionMatrix(confusion, footer = "")
}
\arguments{
\item{confusion}{An object of class \code{\link{ConfusionMatrix}}.}

\item{footer}{The text to display as the footer of the output.}
}
\description{
\code{PrintConfusionMatrix}
}
\details{
Displays a confusion matrix as a heatmap.
}

