% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{probabilities}
\alias{probabilities}
\title{\code{probabilities}}
\usage{
probabilities(object)
}
\arguments{
\item{object}{A model of some kind.}
}
\description{
\code{probabilities}
}
\details{
Computes probabilities that are applicable from the relevant model. For exmaple, probabilities
of class membership from a refression model. This is included for backwards compatibilty.
}

