% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Probabilities.Regression}
\alias{Probabilities.Regression}
\title{\code{Probabilities.Regression}}
\usage{
Probabilities.Regression(object)
}
\arguments{
\item{object}{A model of some kind.}
}
\description{
\code{Probabilities.Regression}
}
\details{
Computes probabilities that are applicable from the relevant model. For exmaple, probabilities
of class membership from a refression model.
}

