% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{PValueAdjustFDR}
\alias{PValueAdjustFDR}
\title{Adjust P-values for multiple comparisons using FDR}
\usage{
PValueAdjustFDR(p, alpha = 0.05)
}
\arguments{
\item{p}{Vector of unadjusted p-values.}

\item{alpha}{Overall significance level.}
}
\description{
This method gives different values from those outputted by \code{p.adjust} with \code{method = "fdr"}.
   This difference is that our method uses a single scaling factor to multiply the p-values. However, conclusions
   (i.e. significant or not) does not change.
}
