% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressiondiagnostics.R
\name{residualPlots.Regression}
\alias{residualPlots.Regression}
\title{residualPlots.Regression}
\usage{
\method{residualPlots}{Regression}(model, ...)
}
\arguments{
\item{model}{A \code{\link{Regression}} model.}

\item{...}{Additional parameters to \code{\link{residualPlots}}}
}
\description{
residualPlots.Regression
}

