% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise.R
\name{Stepwise}
\alias{Stepwise}
\title{\code{Stepwise}}
\usage{
Stepwise(object, output = "final", always.include = NULL,
  direction = "backward", steps = 1000)
}
\arguments{
\item{object}{a \code{Regression} object.}

\item{output}{one of \code{final}, \code{detailed} or \code{all}, which determines how much information to display.}

\item{always.include}{a vector of names of variables to always include in the model.}

\item{direction}{one of \code{backward} or \code{forward}, which determines whether to use backward elimination or forward selection of variables.}

\item{steps}{the maximum number of steps to be considered.}
}
\description{
Stepwise selection of predictor variables by AIC.
}

