% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressionsweep.R
\name{summary.RegressionCorrelations}
\alias{summary.RegressionCorrelations}
\title{summary.RegressionCorrelations}
\usage{
\method{summary}{RegressionCorrelations}(object, ...)
}
\arguments{
\item{object}{Regression model.}

\item{...}{Other arugments.}
}
\description{
summary.RegressionCorrelations
}

