% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{vcov.Regression}
\alias{vcov.Regression}
\title{vcov.Regression}
\usage{
\method{vcov}{Regression}(object, robust.se = FALSE)
}
\arguments{
\item{object}{A \code{Regression} model.}

\item{robust.se}{If \code{TRUE}, computes standard errors that are robust
to violations of the assumption of constant variance for linear and Poisson
models, using the HC3 modification of White's (1980) estimator (Long and Ervin,
2000). This parameter is ignored if weights are applied (as weights already
employ a sandwich estimator). Other options are \code{FALSE}, \code{"hc0"},
\code{"hc1"}, \code{"hc2"}, \code{"hc4"}.}
}
\description{
vcov.Regression
}
