% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{weightedSurveyDesign}
\alias{weightedSurveyDesign}
\title{weightedSurveyDesign}
\usage{
weightedSurveyDesign(data, weights)
}
\arguments{
\item{data}{The data frame.}

\item{weights}{Weights The sampling weights.}
}
\description{
weightedSurveyDesign
}

