% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{FitRegression}
\alias{FitRegression}
\title{\code{FitRegression}}
\usage{
FitRegression(.formula, .estimation.data, subset, .weights, type, robust.se,
  ...)
}
\arguments{
\item{.formula}{An object of class \code{\link{formula}} (or one that can be
coerced to that class): a symbolic description of the model to be fitted.
The details of type specification are given under \sQuote{Details}.}

\item{.estimation.data}{A \code{\link{data.frame}}.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process, or, the name of a variable in \code{data}. It
may not be an expression. \code{subset} may not}

\item{.weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{type}{See \link{Regression}.}

\item{robust.se}{Computes standard errors that are robust to violations of
the assumption of constant variance, using the HC1 (degrees of freedom)
modification of White's (1980) estimator (Long and Ervin, 2000). This parameter is ignored
if weights are applied (as weights already employ a sandwich estimator).}

\item{...}{Arguments to the wrapped functions.}
}
\description{
Fits a regression model.
}

