% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{FitRegression}
\alias{FitRegression}
\title{\code{FitRegression}}
\usage{
FitRegression(
  .formula,
  .estimation.data,
  subset,
  .weights,
  type,
  robust.se,
  outlier.proportion,
  ...
)
}
\arguments{
\item{.formula}{An object of class \code{\link{formula}} (or one that can be
coerced to that class): a symbolic description of the model to be fitted.
The details of type specification are given under \sQuote{Details}.}

\item{.estimation.data}{A \code{\link{data.frame}}.}

\item{subset}{Not used.}

\item{.weights}{An optional vector of sampling weights, or, the name or, the
name of a variable in \code{data}. It may not be an expression.}

\item{type}{See \link{Regression}.}

\item{robust.se}{Computes standard errors that are robust to violations of
the assumption of constant variance, using the HC1 (degrees of freedom)
modification of White's (1980) estimator (Long and Ervin, 2000). This parameter is ignored
if weights are applied (as weights already employ a sandwich estimator).}

\item{outlier.proportion}{A single numeric value that determines the percentage of data points to remove from the
analysis. The data points removed correspond to those in the proportion with the largest residuals.
A value of 0 or NULL would denote no points are removed. A value x, with 0 < x < 0.5 (not inclusive) would
denote that a percentage between none and 50\% of the data points are removed.}

\item{...}{Arguments to the wrapped functions.}
}
\description{
Fits a regression model.
}
