% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{notValidForPartial}
\alias{notValidForPartial}
\title{notValidForPartial}
\usage{
notValidForPartial(object, method)
}
\arguments{
\item{object}{A Regression object}

\item{method}{The regression method.}
}
\description{
notValidForPartial
}

