% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Observed.Regression}
\alias{Observed.Regression}
\title{\code{Observed.Regression}}
\usage{
Observed.Regression(x)
}
\arguments{
\item{x}{A 'Regression' model.}
}
\description{
Observed values used in fitting a model with an outcome variable.
}

