% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{PrintConfusionMatrix}
\alias{PrintConfusionMatrix}
\title{\code{PrintConfusionMatrix}}
\usage{
PrintConfusionMatrix(x)
}
\arguments{
\item{x}{A fitted model.}
}
\description{
\code{PrintConfusionMatrix}
}
\details{
Displays a confusion matrix as a heatmap.
}

