% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Probabilities.Regression}
\alias{Probabilities.Regression}
\title{\code{Probabilities.Regression}}
\usage{
\method{Probabilities}{Regression}(object, newdata, ...)
}
\arguments{
\item{object}{A \code{Regression} object.}

\item{newdata}{Optionally, a data frame including the variables used to fit the model.
If omitted, \code{object$model} is used after any filtering.}

\item{...}{Additional arguments (not used).}
}
\description{
\code{Probabilities.Regression}
}
\details{
Computes probabilities that are applicable from the relevant model. For exmaple, probabilities
of class membership from a regression model.
}
