% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise.R
\name{Stepwise}
\alias{Stepwise}
\title{\code{Stepwise}}
\usage{
Stepwise(object, output = "Final", direction = "Backward",
  always.include = NULL, steps = 1000)
}
\arguments{
\item{object}{a \code{Regression} object.}

\item{output}{one of \code{Final}, \code{Detailed} or \code{All}, which determines how much information to display.}

\item{direction}{one of \code{Backward} (backward elimination of variables) or \code{Forward} (forward selection of variables).}

\item{always.include}{a vector of names of variables to always include in the model.}

\item{steps}{the maximum number of steps to be considered.}
}
\description{
Stepwise selection of predictor variables by AIC.
}

